remove global BuiltinTypes, move type construction to TypeEnvironment
Summary:
This diff completely removes BuiltinTypes and merges its functionality with the TypeEnvironment. This change shifts the control of type creation to the type environment, and we no longer have global types anywhere.

Instead, we thread through the type_env to classes, so that types can reference the env they were created in. We also remove the global BUILTIN_TYPES and instead start constructing type environments as necessary for tests and the compiler.

Next up: Handling exact_types through the type environment for builtin types.

Reviewed By: carljm, wmeehan

Differential Revision: D34082515

fbshipit-source-id: 77f13e2572