always use SEQUENCE_{GET,SET} for exact lists
Summary:
If we have `somelist[0]` where `somelist` is an exact list, it'll be faster to unbox the
int and use `SEQUENCE_GET` than go through a normal `BINARY_SUBSCR`. So
let's do that.

This is a re-land of D33549942 (https://github.com/facebookincubator/cinder/commit/a989f43884f0418d4c117c215c38a5f08248f437), with an added fix to adjust negative indices for
`SEQUENCE_SET`, a bug which the added uses of `SEQUENCE_SET` surfaced.

Reviewed By: sinancepel

Differential Revision: D33960228

fbshipit-source-id: c422c63b3b