check return type of patched free funcs in strict modules
Summary:
Fixes a soundness hole in invoking free functions that could be patched.

We enforce return type of invoked patchable functions by wrapping the patched-in
replacement in a thunk that runtime-checks the return type. Previously this wrapping
only occurred via vtables and thus only on methods attached to types. This diff also
uses thunks to wrap free functions patched directly on a strict module, so that we get
this return type enforcement on free functions also.

This adds a new way in which Static Python is only sound if used with strict modules.
In principle we could provide this same behavior with non-strict modules, but it would
be more invasive in upstream code.

Reviewed By: sinancepel

Differential Revision: D33248904

fbshipit-source-id: f3af586ad7