raise is terminal
Summary:
We should be marking `raise` statements as terminal, just like `return`.

It may seem like the terminality of `raise` is different from `return`, since exceptions
can be caught, but in fact they are the same. A `return` may return an expression which
raises an exception when executed, or an earlier statement in the same block as the
`return` may raise an exception so the `return` is never reached. Thus, statements that
can suppress exceptions (i.e. `try` and `with`) already don't propagate terminality (even
of `return`), so we can handle `raise` as equivalent to `return`.

Reviewed By: DinoV

Differential Revision: D34096378

fbshipit-source-id: d57e9d7e8d