allow unboxing double from int
Summary:
Since int should be treated as a subtype of float, we should allow unboxing a double
from an int.

This also requires fixing a wrong assumption in the JIT that CAST
opcode can only narrow (not change) the JIT type of its operand; this
has not been true since a while back we allowed CAST to cast int to
float, but apparently it didn't cause any test to break until this diff.

Also added a test for unboxing a double from dynamic, which already worked correctly
but we didn't have clear coverage for.

Reviewed By: sinancepel

Differential Revision: D34158432

fbshipit-source-id: 4fe0499324