support literal bool annotations and runtime cast
Summary:
Support annotations using `Literal[...]` syntax. We can only resolve this to a real type
if that type supports runtime casts, otherwise we have to resolve it to dynamic. For now
only support `Literal[True]` and `Literal[False]`, though it will be easy to add literal int
support too, so I'll probably go ahead and do that in the next diff.

Runtime casts for literal bools are supported not via the `CAST` opcode and type
descrs, which would require significant classloader changes, but rather via emitting
bytecode to perform an identity test and raise the appropriate `TypeError` if the test
fails.

One consequence of not having type descr support for literal types is that we can't
`REFINE_TYPE` to a literal type, but I think this shouldn't matter as in principle the
JIT should be able to derive the full literal type information from the identity check
and terminal branch anyway.

Reviewed By: sinancepel

Differential Revision: D34147110

fbshipit-source-id: 6ca0855d9f