support literal int annotations
Summary:
Support annotations in the form `Literal[1]` and runtime checking against them.

This doesn't have an immediate use case (unlike literal bools), but it just seems odd
to support it for bools and not for ints.

This required teasing out the concepts of exactness and literalness a bit more clearly.
It's not really correct to say that `int` is the non-exact version of `Literal[1]`; really the
literal type is inherently exact (aka final) and thus has no inexact version (a literal 1
has no subtypes). If you want to transform a literal type to its non-literal base type, we
should explicitly use `.nonliteral()` for that.

Also we had an override of `bind_constant` in `NumInstance` that served only to
prevent us from inferring int literal types when we otherwise could, so just remove that.

Reviewed By: sinancepel

Differential Revision: D34150114

fbshipit-source-id: 173c28055d