Introduce an API for profilers to collect stacktraces
Summary:
Profiling code needs access to Python level stack traces.

This diff introduces a new function which populates given arrays with
pointers to the code-objects (one array for the async stack, the other
for the sync stack). Additionally, we also accept and populate line
number arrays for both sync and async stacks.

Async stack is needed to figure out the exact code path being hit. The
sync stack is required so that we can build a mixed stack.

Reviewed By: mpage

Differential Revision: D33432728

fbshipit-source-id: 799648c674