Basic support for putting code on huge pages
Summary:
This diff introduces our own `CodeAllocator` type which can wrap around AsmJIT's `Runtime` class or our own allocator which places all JIT code onto huge pages (while available).

The implementation is a very basic grow-only allocator: we keep fillling up chunks of memory allocated 2MB at a time, if a new piece of code doesn't fit in the space remaining in the current chunk we allocate another. When allocating chunks we first try to allocate using huge-pages, but if this fails we fall back to using multiple pages of the default size.

The biggest limitations of this allocator are:
1. No facility to reclaim memory in partially filled chunks (when a new function doesn't fit we start a new chunk).
2. No ability to free memory.
3. We either do or do not use a huge page for a chunk. I.e. it doesn't use Linux's "Transparent Hugepage" support which is more of a hint to the OS we might want these pages to be huge.

With Instagram's workload I've observed ~2.5% of memory allocated for JIT code is wasted due to (1), which doesn't seem too bad for now. I've gone with (3) because it's easier to understand and we seem to have enough free huge pages on our servers for now.

Overall the performance improvement from this change seems to be about 0.6%.

Reviewed By: mpage

Differential Revision: D33834913

fbshipit-source-id: ceef3ab917