backport Annotated changes to typing.py from upstream 3.10 to Cinder's typing.py
Summary:
Context: We'd like to be able to use `Annotated` for user-defined exact types, but having these only live in typing_extensions
disallows us from doing so. `typing_extensions` checks for the existence of Annotated in typing before defining it regardless, so putting in this
stopgap measure should be invisible to users and allow us to export annotated types in the meanwhile. Once we upgrade to 3.10, this should go away completely.

Reviewed By: carljm

Differential Revision: D34243192

fbshipit-source-id: e01272c8e6