Process polymorphic operands from type_profile
Summary:
If we want to start storing all types seen from a profile for all
instructions, we need to store `vector`s of `vector`s to account for all
the type combinations seen for instructions that take multiple operands.

Since this changes our serialization format, we also need to bump the
version number up.

Since we still want to emit `GuardType`s for now, the `HIRBuilder` can
just special-case monomorphic profiles, and stop emitting `GuardType`s
once the rest of the `Specialization` infrastructure is complete.

Reviewed By: carljm

Differential Revision: D34143995

fbshipit-source-id: 4b36d39e02