Add DominatorAnalysis
Summary:
Add a `DominatorAnalysis` that calculates the immediate dominator of
each `BasicBlock` in a `Function` upon initialization. The dominator
information for each block can be queried through a method on the class.

The dominatior analysis algorithm was based off of HHVM's
implementation, which itself was based off of Cooper, Harvey, and
Kennedy's "A Simple, Fast Dominance Algorithm".

Reviewed By: mpage

Differential Revision: D34247629

fbshipit-source-id: a058ff1281