backport Annotated changes to typing.py from upstream 3.10 to Cinder's typing.py
Summary:
This is a resubmission of backporting `typing.Annotated` from 3.10 as well as the 3.10 version of `get_type_hints`.

We need to change both, as typing_extensions provides a get_type_hints if Annotated is not defined - if Annotated is defined in typing, get_type_hints is expected to support stripping/preserving annotations. Client code relies on this.

Reviewed By: carljm

Differential Revision: D34390723

fbshipit-source-id: 3491bd7cd9