Systematise JIT flag parsing
Summary:
This change introduces a `FlagProcessor` in `jit_flag_processor.cpp` which is called as part of the initialization process of the JIT to handle -X opt flags specified via the command line or as environment variables. Callbacks and references are passed as options to an instance of `FlagProcessor`, which are invoked and assigned to respectfully, when the appropriate command line or environment variable is defined.

This allows for three things:
 * The consolidation of flag definitions in one place within `Jit/pyjit.cpp` via a declarative API.
 * `jitXOptionHelpMessage` is called on the `FlagProcessor` instance in order to create a nicely formatted textual representation of the JIT flags.
 * A new flag: `-X jit-help` - which when outputs the aforementioned rendering of the JIT flags and exists.

Reviewed By: carljm

Differential Revision: D34198795

fbshipit-source-id: 69e99b3caf