Use FrameState in LIR generation and add RuntimeFrameState
Summary:
If a function is inlined into another function, it will need the right
`code`, `globals`, and `builtins` (not necessarily the one on the current
HIR function).

Inlined functions will also need these outside of deopt (no access to
deopt metadata), so add `RuntimeFrameState` so that we can traverse the
stack in `sys._getframe` and read globals in `PyEval_GetGlobals`.

Rename `PYSF_PYCODE` to `PYSF_RTFS`.

Reviewed By: mpage

Differential Revision: D31209750

fbshipit-source-id: ae1a0ec563