Avoid emitting INVOKE_FUNCTION on cached properties
Summary:
Currently, cached properties rely on thunks in order to support
use-cases such as inheritance+overrides in static/non-static code as
well as patching.

The problem is, the thunks are created "on demand" whenever
needed. The JIT however burns in the addresses of these whenever an
INVOKE_FUNCTION is emitted through a cached property on a Final class.
This causes a use-after-free because the heap allocated thunk is freed
after the pre-loader does its thing.

The correct fix here is to tie down the lifetimes of cached property
thunks with the lifetime of a v-table. (As an aside, we currently don't
create v-tables when invoking methods on Final classes through
INVOKE_FUNCTION, but that's a separate fix which also needs to happen).

Unfortunately, the above fix is kind of hard to achieve without major
refactoring on the classloader APIs. A few problems that came up when I
started this work:
- `classloader_get_member()` is called *before* a v-table is initialzed.
Since this is the main function that creates thunks, it means that a
thunk is created even before a v-table exists.
- Trying to solve the above issue by creating a v-table before looking
up members is not simple, because v-table construction itself requires
us to look up members.
- Going the other way, caching thunks in the v-table when they are
created is also possible. However, the result is very ad-hoc (and
introduces a lot of debt in the classloader code). For example, we'd
need separate APIs for looking up members from a module and a type.

Reviewed By: wmeehan

Differential Revision: D34489729

fbshipit-source-id: ab87b06232