docs: add GH button in support of Ukraine (#70)
Summary:
Our mission at [Meta Open Source](https://opensource.facebook.com/) is to empower communities through open source, and we believe that it means building a welcoming and safe environment for all. As a part of this work, we are adding this banner in support for Ukraine during this crisis.

Pull Request resolved: https://github.com/facebookincubator/cinder/pull/70

Reviewed By: itamaro

Differential Revision: D34585375

Pulled By: dmitryvinn-fb

fbshipit-source-id: 43688c3544