add basic support for StringEnum's
Summary:
This introduces a very simple StringEnum implementation to Static Python. The approach taken is to fall back to dynamic behaviors and lay the groundwork to optimize things later if we'd like.

One tricky behavior is supporting StringEnum's also being string subclasses. To avoid issues with invokes, etc, copy the appropriate methods from the `Enum` class for the `StringEnum` class instead of using multiple inheritance.

Reviewed By: carljm

Differential Revision: D34478225

fbshipit-source-id: ac6e27cd80