Fix crash in shadowcode when loading from type w/o attribute
Summary: Currently we crash when attempting to load an attribute from a type after we've already created an entry for an instance, but the attribute doesn't exist on the type.  This causes us to check for the value when initializing the cache, and force a cache miss when it's not there.  The end result is that we raise an AttributeError instead of seg faulting.

Reviewed By: mpage

Differential Revision: D34561347

fbshipit-source-id: f964a1291c