support infinite while loops
Summary:
Extend while-loop type inference to know that if the loop condition can
never be falsy (i.e. its type is a truthy literal type), and the loop never breaks,
then the loop itself is terminal (i.e. code after it is unreachable.) This is
necessary to keep existing code examples with infinite loops compiling once
we add support for validating implicit None return at the end of functions.

Reviewed By: sinancepel

Differential Revision: D34554004

fbshipit-source-id: b148ef58e8