fix patching of free func in source module
Summary:
D33248904 (https://github.com/facebookincubator/cinder/commit/e3cb629a91d2e71ccec2eec9ee04d6846a7a3c51) was broken for the case where a function is patched in its original module X
before being imported to another module Y. In this case we resolve the invoke target in module Y,
but module Y has no clue that a patch ever occurred. In the regular interpreter this just meant
we'd fail to enforce the statically expected return type. In the JIT or shadowcode, where we
rely on getting a thunk to use as an indirection layer when patching is enabled, we would
fail to resolve the static return type of the faux "original" function (since it's patched in and
not a static function) and crash.

To fix this, when resolving an imported function we have to trace it back through the
chain of imports to the module where it was originally defined, and check if it was patched
anywhere along that chain before we imported it.

To make that possible, the compiler records imported-from metadata in the last entry of
`co_consts` in the module code object. If we have `from modx import orig as func` in
`mody`, we'll record `"func": ("modx", "orig")` in the imported-from metadata of `mody`.
(This is tracked as a dictionary in the compiler and reconstructed as a dictionary in the
StrictModule object, but passed through `co_consts` as a tuple of 2-tuples since dictionaries
can't be consts.)

The `originals` dict of a StrictModule used to be populated with a full copy of the initial module
contents, but only when that module was patched. Now we populate it granularly by name, and
we populate it for a given name (throughout the import chain from the module where we are
resolving it back to its module of definition) as soon as we resolve that name, even if no patching
has happened. We have to lookup through the module chain to get the original value on any
function resolution, since the immediate container module may not know about a patch that
affected it. Eagerly updating `originals` dict all along the chain makes subsequent lookups
faster. The invariant maintained by the `originals` dict should be that it only ever stores _real_
originals that we've validated all the way back to module of definition.

Reviewed By: DinoV

Differential Revision: D34645658

fbshipit-source-id: f415271015