Make static Enum non-primitive by default
Summary:
In order to accelerate adoption, we need a `__static__.Enum`
that works as a drop-in replacement for `enum.Enum`.

Primitive values cannot be used as `dict` keys or values.
As this is a common use case for enumerations,
we must make `__static__.Enum` non-primitive.

Rename the user-visible `CEnumType` to `__static__.Int64Enum`,
and introduce a new `EnumType` to the compiler that behaves
like vanilla Python enums but with statically-known values.

This allows us to make `StringEnum` subclass `Enum` using mixins.

Reviewed By: DinoV

Differential Revision: D34768300

fbshipit-source-id: 1f36677c88a243bc2a82fee78bdae011a257daac