Finish implementing the type binder
Summary:
With this, everything that declares a name is now tracked, and errors are generated if unknown names are encountered.

This means support for yield expressions, list, dict, and set comprehensions, if exprs, bin, bool, unary, compare, await, subscript expressions, generator expressions, named expressions, as well as yield from, try..catch, with, and async with statements, along with tracking names brought in via imports. Support for destructuring assignments is also added, with special support for assignments of equal size to appropriately propagate the readonlyness of the right hand side of the statement.

Inference for attributes, tuples, lists, sets and dicts is also supported.

Notably, yield from will not be supported due to complexities in how it's used. As it's a rarely used feature and can always be rewritten as `for i in b: yield i`, not loss of usability is expected.

This also makes the readonly test classes plural with the trailing s to align with the static and strict tests.

Differential Revision: D34466897

fbshipit-source-id: 38fac741f342d7dea1ad19aa2ca31078efd14839