understand context manager __exit__ returns
Summary:
To avoid every method decorated with `ContextDecorator` being flagged as
possibly returning implicit `None`, we need a way to signal that it does not ever
suppress exceptions (since `ContextDecorator` wraps the entire body of the
method in a context manager -- if it suppressed exceptions, then any exception
in the method body would result in returning implicit `None`.)

Context managers suppress exceptions by returning `True` from their `__exit__`
method, so a logical way to mark a context manager as never suppressing is
to annotate its `__exit__` method with `-> Literal[False]`. Then type check will
ensure it does not ever return `True`.

This diff adds detection of `-> Literal[False]` annotation on `__exit__`, and if found
will trust that context manager cannot suppress, so if its body is terminal it can
also be marked as terminal.

This requires the ability to resolve an arbitrary attribute name on a type. Since
all of our resolution code depends on passing around an `ast.Attribute`, this is
a bit tricky. I looked into changing all the `resolve_attr_*` APIs to accept a base
node and a string attribute name instead, but we also sometimes use the context
(`Load` vs `Store`), so that ends up requiring three new arguments everywhere;
kinda heavy. So instead I went for just synthesizing an `ast.Attribute` node to
support this. To make this safer, I avoided storing the full `Attribute` node on
`MethodType` and similar results of descriptor resolution, instead storing just
the base expression, which is all any of them actually use anyway.

The resolve methods also previously accepted a `ReferenceVisitor`, but here we
need to resolve an attribute from the type binder. All of our uses of `visitor` rely
only on API that is available on `GenericVisitor`, so I changed all the resolve
methods to accept any `GenericVisitor`.

Reviewed By: sinancepel

Differential Revision: D34829038

fbshipit-source-id: 63fd0f5f0359f70b64d72dd98dd1bed9bf6226c1