Don't copy FrameHeader into generator suspend data
Summary:
For now the `FrameHeader` space for JIT functions contains an on-stack shadow frame. JIT generators keep their shadow frame in their generator object.

This diff stops the initial-yield for a generator from wasting time copying over this FrameHeader data.

Note space is still allocated in the generator object for this data - it's now just unfilled. I don't think there is a way to get around this as we need the relative locations of spills from `RBP` to be constant. This would be particularly hard to work around with eager execution of coroutines.

Reviewed By: tiansihu

Differential Revision: D34360791

fbshipit-source-id: 2d6121ede2b14c1fa4000f9eed6ce115d7d7ac3f