invoke support for typed descriptor slots with defaults
Summary:
This diff introduces vtable support for `typed_descriptor_with_default_value`s. The method of doing so is similar to the strategy we took for properties - when a typed descriptor with default value is detected, we also create entries for the getter and setters. We invoke them using the `(name, "tdget"/"tdset")` form as the last element of the descriptors.

Additionally, the diff introduces inheritance support for typed descriptor slots with defaults. The way I chose to implement this is by creating a new `_static` function to inherit descriptors by copying their offset and creating a new default. One limitation of this is that we don't support writing to overrides of typed descriptors with default on non-static subclasses.

Reviewed By: czardoz

Differential Revision: D34557100

fbshipit-source-id: 4100499eb9cf392daea164c384bb2f0e9980bbd4