fix Class.instance_name[_with_exact]
Summary:
In cases where e.g. `BoolClass.instance_name` was used in a message,
it would show up as `bool` instead of `Literal[False]` even if it was actually
the latter. This is because the literal handling happens in `BoolInstance.name`
but `Class.instance_name` was hardcoded to the class qualname and
didn't actually defer to the instance name. This diff fixes that.

Reviewed By: sinancepel

Differential Revision: D34953726

fbshipit-source-id: a071172341e084e488dfb050bb7301a070038d94