introduce a verification pass to the LIR
Summary:
Context: For the future planned code layout work, we will start potentially emitting physically
adjacent blocks in the LIR to different sections, in which case a jump will need to be inserted.
This will mean that the current LIR invariant of assuming that the block that comes after is a fallthrough
will be incorrect if the code layout algorithm chooses to place them in different sections.

To lay the groundwork for this future planned work, implement a verification pass which verifies current LIR assumptions.
I plan to modify this pass to disallow fallthroughs to blocks in different sections in the future.

Code layout/branch insertion will only happen after register allocation, which is why this pass is
only meant to check those invariants. We might also want to check a different set of invariants
pre-register allocation. I chose to keep things simple here and not go with the generic
design/multiple functions for now, as I don't know what we'll want to share between those yet.

Reviewed By: mpage

Differential Revision: D34873159

fbshipit-source-id: e936dc8f40ce597c0b2a5dfe3728e6a031540386