introduce code section hints to LIR basic blocks and add parsing support
Summary:
This diff adds basic code section support for LIR blocks, and of their parsing (for testing purposes). The hints are unused for now, future diffs will start emitting code in different section based on hints, and start using basic heuristics to determine whether a basic block should be hot or cold on the LIR side.

Since there is a potential future in which we'll use HIR-level hints for basic block hotness, the CodeSection enum lives in `codegen` instead of being attached to LIR.

For now, I decided to omit code sections from printing of LIR blocks. I plan to start printing them in the same diff where we start using these hints.

Reviewed By: mpage

Differential Revision: D34900318

fbshipit-source-id: aedd6614e55507192a2b9d42d2a1dd8acac1ad6e