Convert narrowed isinstance calls into limited type checks
Summary:
This converts our `isinstance` calls into a `CAST` opcode where the result is either True or False.  This limits the type checks to only supporting normal subtype relationships and not dynamic behavior such as proxys defining `__class__`, or `__instancecheck__`.  Rather than introducing a new opcode this just makes the existing opcode take a new boolean parameter indicating whether it's error reporting or not.

Currently this means we'll lose our normal `isinstance` optimization of avoiding a boxed bool and truthy check in the JIT, but that'll get resolved in the next diff.

Reviewed By: carljm

Differential Revision: D33268361

fbshipit-source-id: 5c2d8473fe3bd4483ebd206a545c2e4ac2db2648