Unified format for immutable errors
Summary:
Modify the immutable error reporting API so it's easier to report an error from the cinder runtime.
Under this new API, error reporting should we pre-defined error categories along with arguments.

To define a new error category, create a new `PyImmutableErrorEntry` in `errors.c` (TBD might want a separate file for this), of the form `{err_code, err_format, nargs}`. `err_code` is a unique number identifying the error. `err_format` is a cpython format string, and all its argument must expect `PyObject*`. `nargs` is the number of arguments expected by the format string.

To report an error, use the `_PyErr_IMMUTABLE_ERR` macro. The first argument is the pointer to the error category defined above. The next arguments are arguments to the format string. This macro will add both the formatted string and the arguments as part of the final error object, which the python side handler can use.

Next up: add stack traces to the error object.

Reviewed By: tiansihu

Differential Revision: D34657350

fbshipit-source-id: 9fd7fa3bf716a7c71141d3752425aee7f70b955e