update ContextDecorator for implicit None returns
Summary:
ContextDecorator rewrites the entire body of the decorated function to be wrapped
in a context manager. Since context managers by default must be assumed to
possibly suppress exceptions, this means every ContextDecorator-wrapped function
might return `None`, and every caller would have to account for that. This isn't
workable in practice; too many ContextDecorator subclasses never suppress
and callers of their decorated functions can't all be required to handle `None`.

We have the ability to mark context managers as never suppressing exceptions by
having their `__exit__` method return `Literal[False]`. Update `ContextManager` so
that it does this, and inherits from a base `ExcContextManager` which does not do
this and can still suppress exceptions.

This also requires adding a rudimentary `SelfType`, currently usable only for builtin
methods, so a method can understand that it always returns the precise type of
`self`, even in subclasses. This is necessary so we know that
`ContextManager()._recreate_cm()` returns a `ContextManager`, even though it
only exists on the base `ExcContextManager`. There's a PEP for an actual `SelfType`
annotation to be included in `typing` module, so in future we can adapt this support
to cover user annotations, too.

In keeping with how decorator typing usually works in Python, our approach to
typing for `ExcContextDecorator` decorated functions is that the return annotation
does not need to include `| None` if the visible body of the function cannot return
`None`. Even though the implementation modifies the internal body of the function,
from a type annotations perspective we retain the idea that the `None` return
is added externally by the decorator. So we do this:

```
MyExcDecorator()
def f() -> int:
    return 42
```
 And the decorator is understood to turn the `-> int` function into an `-> int | None`
function.

Reviewed By: DinoV

Differential Revision: D34970297

fbshipit-source-id: d6968aeb5063ef9524899cda6d70138b7151dd24