Resolve decorators after class body
Summary:
For `dataclass`, we will need access to resolved members in order to generate method code.

To fix, we should resolve decorators only once the entire class body has been visited.

This has not been an issue before because existing decorators do not depend on any of the class members.

Reviewed By: carljm

Differential Revision: D35257494

fbshipit-source-id: 082082704fbf6b9242f0b8a494ae76463b5ca209