Add _PyDict_HasUnsafeKeys()
Summary:
This returns true if the given dict has any deferred values, or any
keys that are not exactly `str` (it also returns true if anyone has even looked
up in the dict with a non-`str` key; I'm hoping that isn't a problem for our
use case).

The target use case is for looking up items in the JIT, where we need to avoid
Python heap mutations (including arbitrary code execution).

Reviewed By: carljm

Differential Revision:
D35732041

LaMa Project: L1125187

fbshipit-source-id: 6f2c739fec236e10df3a447646a05ea69aa1547d