add (sometimes dynamic) types for some builtin and typing objects
Summary: The goal here is to support erroring on unknown names. To accomplish that, we'll need to have a more comprehensive understanding of builtins in the Static Compiler. Towards that end, add typing information to cover the cases overseen in our unit tests (my methodology was to add a print statement to visitName to see when we failed to locate a name in a scope/resolve it).

Reviewed By: DinoV

Differential Revision: D35127137

fbshipit-source-id: be9b2ee440b257c03be936633cd12f6b7ce9214d