Port freelist size changes
Summary:
Port Cinder changes to dict and list freelist sizes.

Combined bits from D22982625 (dict) and D23285552 (list).

These are now wrapped in `ifndef` which means that instead of landing
this diff, we could also leave the code as-is and set them in
configure via `CFLAGS`, e.g. `-DPyList_MAXFREELIST=200`.

Reviewed By: itamaro

Differential Revision: D35331800

fbshipit-source-id: 14ba496d9c15a9cdb74d6ba600646d110ce9358b