Clean up class decorator resolution
Summary:
An earlier version moved decorator resolution to `finish_bind`,
so there was a worry of existing references to the class having members
in the case of a decorator resolving the type to `dynamic`.

Since we still resolve decorators in `visitClassDef`,
we can safely drop the class on the floor rather than clearing its members.

Reviewed By: carljm

Differential Revision: D35331585

fbshipit-source-id: b466827a24995dae57e197d64baf959cbf253500