Add return type annotations for some str methods
Summary:
This diff adds support for annotating return types of reflected methods. For
example, `str.upper` is currently untyped. However, we know that no matter the
caller, it will always return an exact `str`.

It also adds support for annotating that a method is still valid to use on
arbitrary subclasses --- if we don't know which subclass it is statically.

This all means, overall, fewer CAST opcodes emitted.

Reviewed By: carljm

Differential Revision: D35297003

fbshipit-source-id: 217028137e01247c117b436eaa634a078719ba25