Remove BeginInlinedFunction and EndInlinedFunction in trivial cases
Summary:
When a function is just returning a constant, for example, we
can get rid of the shadow frame bookkeeping. It is not needed for deopt
or otherwise reifying frames. Such a function will likely not even show
up in profiles, so don't bother keeping a shadow frame around for it.

Reviewed By: swtaarrs

Differential Revision: D35374703

fbshipit-source-id: 3806bab19501d37b6a677cff63c6923ebe9293bb