add an explicit type for overloaded functions
Summary:
Context: For supporting erroring on unknown names, we will need to start explicitly registering UnknownDecoratorMethod's as dynamic. We currently remove them completely from a module/class' local types. The reason we've refrained from adding UnknownDecoratorMethods is another check: We error if a user attempts to declare a local with the same name as a previous one, and this following pattern would do that, assuming that overload  is something the static compiler doesn't understand:

```
overload
def foo(x: Literal[3]) -> Literal[3]: ...
def foo(x: int) -> int: ... # Declaring over another local!
```

To set ourselves up for reintroducing these declarations in the future, instead create an OverloadedFunction class, and manually stop declaring these in the function group.

This cropped up when trying to error on unknown names - we don't have the type context
yet when we visit the for target, that happens in the code after that visit. It looks like
we can safely eliminate this visit here since `assign_value` sets the type of the node.

Reviewed By: wmeehan

Differential Revision: D35315954

fbshipit-source-id: b6219e97e43dae647bb361b329025714d2d17c2b