refactor static tests in preparation for erroring on unknown names
Summary:
Since we didn't mind unknown names before, we were liberal in the uses of variables. Now that we'll actually error, let's refactor our tests to define the names they use.

The most significant logical change is that `bind_stmt/expr` no longer requires that the body of the test has length 1, but instead operates on the last statement/expr.

Reviewed By: carljm

Differential Revision: D35317649

fbshipit-source-id: dc5a4d612dfdf8fc049920140f3bc4db1cea58b2