understand property setters and start declaring all decorated functions
Summary:
We previously had the behavior of not declaring any locals which corresponded to decorated functions if we don't understand decorators. This behavior helped us support `foo.setter` and `overload` behavior without having to understand these decorators which don't have a runtime effect. However, now that we're looking to error on unknown names, we need to change this behavior.

Towards that end, we introduce the concept of a `TransientDecoratedMethod` which indicates that the decorated method will go away in the runtime. Any typing.overload functions and function setters, on successful decorator application, will set the decorated method's type to this, and the rest of the machinery will avoid declaring these methods.

To support property setters, there was one more nuance: We resolve decorators in the decl visitor, where we don't have access to type binder data. This is a problem when trying to declare property function groups, as the `foo.setter` decorator depends on understanding what `foo` is, which is declared in the same function group. To help with this, we introduce a way to add type annotation overrides to reference visitors, and add overrides for known decorated functions when finishing binds in a function group.

Reviewed By: wmeehan

Differential Revision: D35324102

fbshipit-source-id: de1f385ec11b487719ac01fdbc79c0c052e1ef39