clean up incompatible cast warning
Summary:
I missed a compiler warning on incompatible cast here, since the
fastcall signature is not the same as the varargs signature. This seems
pretty ugly, but looking around the codebase it seems the explicit cast
to `void(*)(void)` is how this is handled elsewhere, so let's follow the
same pattern here.

Reviewed By: itamaro

Differential Revision: D35422391

fbshipit-source-id: 4545bccedd8006a7f5b369ceecc4162b82b50dd9